requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.6.4
pdf2image==1.17.0
Pillow==10.4.0
